refactor(home): extract auth bootstrap logic from HomePage
